refactor: replace SliverAppBar with PortfolioSliverAppBar for improved structure and navigation
