+++ lib/pages/home_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../widgets/header_widget.dart';
 import '../widgets/hero_section.dart';
 import '../widgets/about_section.dart';
 import '../widgets/skills_section.dart';
@@ -27,33 +28,7 @@ class HomePage extends StatelessWidget {
       backgroundColor: Colors.white,
       body: NestedScrollView(
         headerSliverBuilder: (context, innerBoxIsScrolled) => [
-          SliverAppBar(
-            backgroundColor: Colors.white,
-            elevation: 1,
-            shadowColor: Colors.grey.withAlpha(25),
-            toolbarHeight: 80,
-            automaticallyImplyLeading: false,
-            floating: true, // Shows when scrolling up
-            snap: true,     // Snaps to show/hide
-            pinned: false,  // Doesn't stay pinned at top
-            title: Text(
-              'Dzikran Azka Sajidan',
-              style: TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.w600,
-                color: const Color(0xFF202124),
-              ),
-            ),
-            actions: [
-              // Navigation items
-              _buildNavItem('About'),
-              _buildNavItem('Projects'),
-              _buildNavItem('Skills'),
-              _buildNavItem('Experience'),
-              _buildNavItem('Contact'),
-              const SizedBox(width: 24), // Right padding
-            ],
-          ),
+          const PortfolioSliverAppBar(),
         ],
         body: CustomScrollView(
           slivers: [
@@ -65,29 +40,6 @@ class HomePage extends StatelessWidget {
               ),
             ),
           ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildNavItem(String title) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 8),
-      child: TextButton(
-        onPressed: () {
-          // TODO: Implement navigation
-        },
-        style: TextButton.styleFrom(
-          splashFactory: NoSplash.splashFactory,
-          overlayColor: Colors.transparent,
-        ),
-        child: Text(
-          title,
-          style: const TextStyle(
-            fontSize: 16,
-            fontWeight: FontWeight.w400,
-            color: Color(0xFF5F6368),
-          ),
         ),
       ),
     );
